% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_wrapper.R
\name{cluster_wrapper}
\alias{cluster_wrapper}
\title{Use cluster_wrapper to perform GSEA quickly and efficiently on your differential expression output}
\usage{
cluster_wrapper(
  data,
  gene_col = NULL,
  stat_col = NULL,
  species = "Mus musculus",
  category = "H",
  inputType = "SYMBOL",
  n_max = 10
)
}
\arguments{
\item{data}{An object of either class data.frame or a named numeric vector in one one of two formats. If data is a data.frame, you must supply gene_col and stat_col (see below). If it is a named vector, the names should be the gene names and the vector should be the stat value}

\item{gene_col}{A character string of length one that is the name of the column in data that contains gene symbols. If NULL, will attempt to default to column 1}

\item{stat_col}{A character string of length one that is the name of the column in data that contains the numeric values for ranking, can be logFC, pval*sign of log, etc. If NULL, will attempt to default to column 2}

\item{species}{character string that is the species of interest for GSEA (eg: Mus musculus, Homo sapiens)}

\item{category}{a character string that fits with msigdbr::, tells the program what to use}

\item{inputType}{a character string of length one that tells the naming convention that is used. Default is SYMBOL, acceptable inputs are also ENTREZID and ENSEMBL}

\item{n_max}{a numeric of length one that tells the max number of GSEA plots to return (does not effect the dataframe). Defaults to 10}
}
\value{
Returns an object of class 'list' that contains all GSEA plots of significantly enriched sets based on the given results file. At the moment, this is limited to A vs B comparisons.
}
\description{
\code{cluster_wrapper} will perform GSEA quickly and efficiently on your Differential expression output. All you need is a dataframe of genes and their associated log fold change. Future updates will allow for multi-level GSEA (eg: cluster marker analysis for scRNA)
}
\examples{
cluster_wrapper(data = df, gene_col = "SYMBOL",stat_col = "avg_logFC")
cluster_wrapper(data = df, gene_col = "SYMBOL",stat_col = "avg_logFC", category = "C2")
cluster_wrapper(data = df, species = "Human",gene_col = "SYMBOL",stat_col = "avg_logFC", category = "H")

}
