% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msigdbr_obj_to_gmt_file.R
\name{convert_msigdbr_obj_to_gmt_file}
\alias{convert_msigdbr_obj_to_gmt_file}
\title{This Function writes a .gmt file for the GSEA protocol}
\usage{
convert_msigdbr_obj_to_gmt_file(list_of_sets)
}
\arguments{
\item{list_of_sets}{A named list where each element in the list is an output of msigdbr}
}
\value{
nothing
}
\description{
This Function writes a .gmt file for the GSEA protocol
}
\examples{
C1_df = msigdbr(species = "Mus musculus", category = "C1", subcategory = "")
C2_df = msigdbr(species = "Mus musculus", category = "C2", subcategory = "CGP")
list_of_sets = list(C1 = C1_df,
                   C2 = C2_df)
convert_msigdbr_obj_to_gmt_file(list_of_sets)              
}
