% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM.Heatmaps.R
\name{GSM.Heatmaps}
\alias{GSM.Heatmaps}
\title{This function takes direct GSMGSEA output and transforms it into hierarchically clustered gene vs gene set heatmaps}
\usage{
GSM.Heatmaps(gsea_results, verbose = TRUE, pdftofile = TRUE)
}
\arguments{
\item{gsea_results}{is the direct output from GSEA function in step immediately prior. There is no need for restructuring, just feed it directly into the function as is.}
}
\value{
pdf heatmap plots corresponding to each gsea_results$Results
}
\description{
This function takes direct GSMGSEA output and transforms it into hierarchically clustered gene vs gene set heatmaps
}
\examples{
snyder_outputs is GSEA result
GSM.Heatmaps(snyder_outputs)
}
