% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM.Multilevel.Mouse.Heatmaps.R
\name{GSM.Multilevel.Mouse.Heatmaps}
\alias{GSM.Multilevel.Mouse.Heatmaps}
\title{GENE VS GENE SET HEATMAPS intended for visualization of clustering or dimentionality reduction nalyses of 3 or more samples
This function generates hierarchically clustered Multilevel Heatmaps which display three samples in one concise plot.
This function takes direct GSMGSEA output and transforms it into a unique visualization}
\usage{
GSM.Multilevel.Mouse.Heatmaps(
  gsea_results,
  verbose = TRUE,
  pdftofile = TRUE,
  filePrefix
)
}
\arguments{
\item{gsea_results}{is the direct output from GSMGSEA function in step immediately prior. There is no need for restructuring, just feed it directly into the function as is.}
}
\value{
.pdf files of multi-colored heatmaps displaying 3 experimental conditions/samples/vectors
}
\description{
GENE VS GENE SET HEATMAPS intended for visualization of clustering or dimentionality reduction nalyses of 3 or more samples
This function generates hierarchically clustered Multilevel Heatmaps which display three samples in one concise plot.
This function takes direct GSMGSEA output and transforms it into a unique visualization
}
\examples{
gsea_results is GSEA result
GSM.Multilevel.Mouse.Heatmaps(gsea_results)
}
