% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMGSEA.R
\name{GSMGSEA}
\alias{GSMGSEA}
\title{Modified GSEA function, this function performs fgsea on the results of previous Differential Expression analysis, Dimensional Reductions and/or other appropriate filtering steps.}
\usage{
GSMGSEA(gene_list, theGoFile, pval, verbose = TRUE, onlypos = FALSE)
}
\arguments{
\item{gene_list}{is a named vector of scores where names must be ENTREZID and scores are a preferred number}

\item{theGoFile}{is a .gmt file name generated by convert_msigdbr_obj_to_gmt_file.R}

\item{pval}{is preferred cutoff for analysis}

\item{onlypos}{a logical flag which drops any ES values returned as negative and thus correcting less harshly. Defaults to FALSE. We advise to set onlypos = TRUE when performing GSMGSEA on NMF vectors.}
}
\value{
nothing
}
\description{
Modified GSEA function, this function performs fgsea on the results of previous Differential Expression analysis, Dimensional Reductions and/or other appropriate filtering steps.
}
