% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moff2Seurat.R
\name{Moff2Seurat}
\alias{Moff2Seurat}
\title{Updating formatting of experiment files}
\usage{
Moff2Seurat(object, project = NULL, new_names = 1)
}
\arguments{
\item{object}{An object of class 'list' with at least 3 items: $ex - the expression matrix or dataframe, $sampInfo - the colData, $featInfo - the rowData, and $metadata (optional)}

\item{project}{A character string of length one to be given as the project name to the Seurat object, defaults to the name of object}

\item{new_names}{Either a numeric index or character string of length one of which column in $featInfo to use as rownames}
}
\value{
Returns an object of class 'SeuratObj' that contains all info from inital experiment contained in those 4 list headings,
see ?\code{Seurat} for help with manipulating new object
}
\description{
\code{Moff2Seurat} will convert a standard list format experiment file into a Seurat Object
}
