% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moff2Summarized.R
\name{Moff2Summarized}
\alias{Moff2Summarized}
\title{Updating formatting of experiment files}
\usage{
Moff2Summarized(object)
}
\arguments{
\item{object}{An object of class 'list' with 4 items: $ex - the expression matrix or dataframe, $sampInfo - the colData, $featInfo - the rowData, and $metadata}
}
\value{
Returns an object of class 'SummarizedExperiment' that contains all info from inital experiment contained in those 4 list headings,
see ?\code{SummarizedExperiment} for help with manipulating new object
}
\description{
\code{Moff2Summarized} will convert a standard list format experiment file into a SummarizedExperiment
}
