% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse2Human.R
\name{Mouse2Human}
\alias{Mouse2Human}
\title{Homology Conversion Function converts mouse ENSEMBLE ID's to human ENTREZ ID'd + HGNC and MGI's}
\usage{
Mouse2Human(MouseGenes)
}
\arguments{
\item{MouseGenes}{is a vector of mouse ENSEMBLE ID's}
}
\value{
nothing
}
\description{
Homology Conversion Function converts mouse ENSEMBLE ID's to human ENTREZ ID'd + HGNC and MGI's
}
\examples{
readcounts <- data$ex
MouseGenes <- data$featInfo$ENSEMBL
humanGenes <- Mouse2Human(MouseGenes)
}
