% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseENS2MouseENT.R
\name{MouseENS2MouseENT}
\alias{MouseENS2MouseENT}
\title{Homology Conversion Function convert mouse ENSEMBLE ID's to mouse ENTREZ ID's + MGI's}
\usage{
MouseENS2MouseENT(MouseGenes)
}
\arguments{
\item{MouseGenes}{is a vector of mouse ENSEMBLE ID's}
}
\value{
nothing
}
\description{
Homology Conversion Function convert mouse ENSEMBLE ID's to mouse ENTREZ ID's + MGI's
}
\examples{
readcounts <- data$ex
mouseENS <- data$featInfo$ENSEMBL
mouseENT <- MouseENS2MouseENT(MouseGenes)
}
