% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseENT2MouseSYM.R
\name{MouseENT2MouseSYM}
\alias{MouseENT2MouseSYM}
\title{Homology Conversion Function converts mouse ENTREZID's to mouse SYMBOL}
\usage{
MouseENT2MouseSYM(MouseGenes)
}
\arguments{
\item{MouseGenes}{is a vector of mouse ENTREZID}
}
\value{
nothing
}
\description{
Homology Conversion Function converts mouse ENTREZID's to mouse SYMBOL
}
\examples{
readcounts <- data$ex
MouseGenes <- data$featInfo$ENTREZ
Symbols <- MouseENT2MouseSYM(MouseGenes)
}
