% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseSYM2MouseENT.R
\name{MouseSYM2MouseENT}
\alias{MouseSYM2MouseENT}
\title{Convert mouse ENTREZID's to mouse SYMBOL}
\usage{
MouseSYM2MouseENT(MouseGenes)
}
\arguments{
\item{MouseGenes}{is a vector of mouse ENTREZID}
}
\value{
nothing
}
\description{
Convert mouse ENTREZID's to mouse SYMBOL
}
\examples{
readcounts <- data$ex
MouseGenes <- data$featInfo$ENTREZ
Symbols <- MouseENT2MouseSYM(MouseGenes)
}
