% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareLLB.R
\name{prepareLLB}
\alias{prepareLLB}
\title{This function converts Affymetrix Probe IDs into ENTREZ, SYMBOL and ENSEMBL homology. Accepts and returns a layered data.frame}
\usage{
prepareLLB()
}
\arguments{
\item{LLB}{is a data.frame of rat Affymetrix ID'sT}
}
\value{
a data.frame with corresponding ENTREZ, SYMBOL, ENSEMBL ID's
}
\description{
This function converts Affymetrix Probe IDs into ENTREZ, SYMBOL and ENSEMBL homology. Accepts and returns a layered data.frame
}
\examples{
prepareLLB(LLB)
}
