% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatENT2RatSYM.R
\name{RatENT2RatSYM}
\alias{RatENT2RatSYM}
\title{Homology Conversion Function converts rat ENTREZID's to rat SYMBOL}
\usage{
RatENT2RatSYM(RatGenes)
}
\arguments{
\item{RatGenes}{is a vector of rat ENTREZID}
}
\value{
nothing
}
\description{
Homology Conversion Function converts rat ENTREZID's to rat SYMBOL
}
\examples{
readcounts <- data$ex
RatGenes <- data$featInfo$ENTREZ
Symbols <- RatENT2RatSYM(RatGenes)
}
