% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vignette.R
\name{run_vignette}
\alias{run_vignette}
\title{Wrapper function to facilitate calling running of vignettes when package is installed, not cloned.}
\usage{
run_vignette(format = c("bulk", "NMF", "scRNA"))
}
\arguments{
\item{format}{A character string of length one that specifies which vignette to build. Must be one of "bulk", "NMF", or "scRNA".}
}
\description{
Wrapper function to facilitate calling running of vignettes when package is installed, not cloned.
}
