% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DE_tools_for_concordance_analysis.R
\name{runDESeq2_indFilterActive_2}
\alias{runDESeq2_indFilterActive_2}
\title{This function is similar to runDESeq2_indFilterActive() except that it doesn't 
replace NA p-adj values by 1. It is simply used to compare number of filtered genes}
\usage{
runDESeq2_indFilterActive_2(
  count.dat,
  conditions,
  colors = "",
  path = "",
  specification = ""
)
}
\description{
This function is similar to runDESeq2_indFilterActive() except that it doesn't 
replace NA p-adj values by 1. It is simply used to compare number of filtered genes
}
