% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2_indFilterDeactive.R
\name{runDESeq2_indFilterDeactive}
\alias{runDESeq2_indFilterDeactive}
\title{This function takes raw count data and performs Differential Expression Analysis of sequence count data based on negative binomial distribution
default: independentFiltering = FALSE, cooksCut = .99 quantile of the F(p, m-p)}
\usage{
runDESeq2_indFilterDeactive(
  count.dat,
  conditions,
  colors = "",
  path = "",
  specification = ""
)
}
\arguments{
\item{count.dat}{is the raw count data}

\item{conditions}{is the sample information for DE}
}
\value{
Results of DE analysis (pval, adjusted pval, logFC, avgExpr, stat)
}
\description{
This function takes raw count data and performs Differential Expression Analysis of sequence count data based on negative binomial distribution
default: independentFiltering = FALSE, cooksCut = .99 quantile of the F(p, m-p)
}
