% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DE_tools_for_concordance_analysis.R,
%   R/runedgeR_exact.R
\name{runedgeR_exact}
\alias{runedgeR_exact}
\title{Menu of all the GeneSetMatch tools for Differential Expression Analysis}
\usage{
runedgeR_exact(
  count.dat,
  conditions,
  colors = "",
  path = "",
  specification = ""
)

runedgeR_exact(
  count.dat,
  conditions,
  colors = "",
  path = "",
  specification = ""
)
}
\arguments{
\item{count.dat}{is the raw count data}

\item{conditions}{is the sample information for DE}

\item{list_of_sets}{A named list where each element in the list is an output of msigdbr}
}
\value{
nothing

Results of DE analysis (pval, adjusted pval, logFC, avgExpr, rank)
}
\description{
Menu of all the GeneSetMatch tools for Differential Expression Analysis

This function takes raw count data and performs Differential Expression analysis on specified condition
}
\examples{
C1_df = msigdbr(species = "Mus musculus", category = "C1", subcategory = "")
C2_df = msigdbr(species = "Mus musculus", category = "C2", subcategory = "CGP")
list_of_sets = list(C1 = C1_df,
                   C2 = C2_df)
convert_msigdbr_obj_to_gmt_file(list_of_sets)       
}
