% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runedgeR_glm.R
\name{runedgeR_glm}
\alias{runedgeR_glm}
\title{This function takes raw count data and performs Differential Analysis of sequence count data via a General Linear Modeling analysis on specified condition}
\usage{
runedgeR_glm(count.dat, conditions, colors = "", path = "", specification = "")
}
\arguments{
\item{count.dat}{is the raw count data}

\item{conditions}{is the sample information for DE}
}
\value{
Results of DE analysis (pval, adjusted pval, logFC, avgExpr, rank)
}
\description{
This function takes raw count data and performs Differential Analysis of sequence count data via a General Linear Modeling analysis on specified condition
}
