% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNOISeq.R
\name{runNOISeq}
\alias{runNOISeq}
\title{This function performs analysis on RNA-seq expression data or other similar kind of data. Exploratory plots to evaluate saturation, count distribution, expression per chromosome, type of detected features, features length, etc. Differential expression between two experimental conditions with no parametric assumptions.}
\usage{
runNOISeq(count.dat, conditions, colors = "", path = "", specification = "")
}
\arguments{
\item{count.dat}{is the raw count data}

\item{conditions}{is the sample information for DE}
}
\value{
Results of DE analysis (pval, adjusted pval, logFC, theta, prob.DE)
}
\description{
This function performs analysis on RNA-seq expression data or other similar kind of data. Exploratory plots to evaluate saturation, count distribution, expression per chromosome, type of detected features, features length, etc. Differential expression between two experimental conditions with no parametric assumptions.
}
