% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuasiSeq_QLSpline.R
\name{runQuasiSeq_QLSpline}
\alias{runQuasiSeq_QLSpline}
\title{This function identifies differentially expressed genes in RNA-seq count data using quasi-Poisson or quasi-negative binomial models.}
\usage{
runQuasiSeq_QLSpline(
  count.dat,
  conditions,
  colors = "",
  path = "",
  specification = ""
)
}
\arguments{
\item{count.dat}{is the raw count data}

\item{conditions}{is the sample information for DE}
}
\value{
Results of DE analysis (pval, adjusted pval, logFC, stat)
}
\description{
This function identifies differentially expressed genes in RNA-seq count data using quasi-Poisson or quasi-negative binomial models.
}
