% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBM_heatmap.R
\name{TB.Heatmap}
\alias{TB.Heatmap}
\title{FOR DEBUGGING PURPOSES ONLY
This function generates a Hierarchically clustered Multilevel Heatmaps which displays three samples in one concise plot.
This function takes direct GSEA output and transforms it into a unique visualization}
\usage{
TB.Heatmap(gsea_results)
}
\arguments{
\item{gsea_results}{is the direct output from GSEA function in step immediately prior. There is no need for restructuring, just feed it directly into the function as is.}
}
\value{
multi-colored heatmaps displaying 3 experimental conditions/samples/vectors
}
\description{
FOR DEBUGGING PURPOSES ONLY
This function generates a Hierarchically clustered Multilevel Heatmaps which displays three samples in one concise plot.
This function takes direct GSEA output and transforms it into a unique visualization
}
