% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformedPM.R
\name{transformedPM}
\alias{transformedPM}
\title{Transformed Pattern Markers protocol designed to identify the most highly weighted outputs from NMF, and extract only the genes which strongly associate with a particular pattern, or a linear combination of patterns.}
\usage{
transformedPM(W)
}
\arguments{
\item{W}{is a matrix with 3 vectors, the direct un-normalized output of the preceding NMF computation}
}
\value{
nothing
}
\description{
Transformed Pattern Markers protocol designed to identify the most highly weighted outputs from NMF, and extract only the genes which strongly associate with a particular pattern, or a linear combination of patterns.
}
\examples{
tpm_LLB_NMF <- transformedPM(LLB_NMF)
}
